Move the t1oo permission file to population_permissions subdir
The t1oo.py file predates the categorisation of permissions into
population permissions, dataset permissions and (current unused)
component access permissions which are passed as `analysis_scope`
to the RAP API. In prepration for updating the handling of t1oo
permissions to bring them in line with the way other permissions
are handled, we move the file into the population_permission
subdir, along with ndoo and gp-activations.
