+++ jobserver/permissions/__init__.py
@@ -1,4 +1,5 @@
-from . import sqlrunner, t1oo, validator
+from . import sqlrunner, validator
+from .population_permissions import t1oo
 
 
 __all__ = ["sqlrunner", "t1oo", "validator"]
